pyproject.toml
[tool.poetry]
name = "moodlewithselenium"
version = "0.1.0"
description = ""
authors = ["ragatzino <antoinebrunetti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.18.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"